pseudo modularizing combinatorial
