+++ include/eve/module/combinatorial/regular/impl/lcm.hpp
@@ -7,18 +7,9 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/value.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/detail/skeleton_calls.hpp>
-#include <eve/function/abs.hpp>
-#include <eve/function/all.hpp>
-#include <eve/function/converter.hpp>
-#include <eve/function/gcd.hpp>
-#include <eve/function/if_else.hpp>
-#include <eve/function/is_flint.hpp>
+#include <eve/module/arithmetic.hpp>
+#include <eve/module/combinatorial/regular/gcd.hpp>
 
-#include <type_traits>
-#include <eve/detail/apply_over.hpp>
 
 namespace eve::detail
 {
