+++ include/eve/module/combinatorial/regular/impl/lfactorial.hpp
@@ -7,20 +7,10 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/value.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/detail/skeleton_calls.hpp>
-#include <eve/function/all.hpp>
-#include <eve/function/factorial.hpp>
-#include <eve/function/inc.hpp>
-#include <eve/function/converter.hpp>
-#include <eve/function/if_else.hpp>
-#include <eve/function/is_flint.hpp>
-#include <eve/function/is_gez.hpp>
-#include <eve/function/lgamma.hpp>
-
-#include <type_traits>
-#include <eve/detail/apply_over.hpp>
+#include <eve/module/arithmetic.hpp>
+//#include <eve/module/special.hpp>
+#include <eve/module/combinatorial/regular/factorial.hpp>
+#include <eve/module/special/regular/lgamma.hpp>
 
 namespace eve::detail
 {
