+++ include/eve/module/combinatorial/regular/impl/nth_prime.hpp
@@ -7,18 +7,7 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/value.hpp>
-#include <eve/detail/apply_over.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/detail/skeleton_calls.hpp>
-#include <eve/function/add.hpp>
-#include <eve/function/gather.hpp>
-#include <eve/function/convert.hpp>
-#include <eve/function/converter.hpp>
-#include <eve/function/if_else.hpp>
-#include <eve/function/inc.hpp>
-#include <eve/function/is_eqz.hpp>
-#include <type_traits>
+#include <eve/module/arithmetic.hpp>
 
 namespace eve::detail
 {
