+++ include/eve/module/combinatorial/regular/impl/prime_floor.hpp
@@ -7,16 +7,8 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/value.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/detail/skeleton_calls.hpp>
-#include <eve/detail/apply_over.hpp>
-#include <eve/function/any.hpp>
-#include <eve/function/average.hpp>
-#include <eve/function/convert.hpp>
-#include <eve/function/converter.hpp>
-#include <eve/function/inc.hpp>
-#include <eve/function/nth_prime.hpp>
+#include <eve/module/arithmetic.hpp>
+#include <eve/module/combinatorial/regular/nth_prime.hpp>
 #include <eve/function/if_else.hpp>
 #include <type_traits>
